move `uncompress2` -> `uncompress`
